test: add re-usable ClusterTopologyDomain for propery tests
